Refactor DialogViewPage
Split component - add 2 additional components for sending messages and message itself
