+++ Frontend/src/Components/Message/Message.jsx
@@ -1,0 +1,3 @@
+export default function Message({ message }) {
+    return <li key={message.id}>Message: {message.content}</li>;
+}
